+++ Services/Twilio.php
@@ -15,15 +15,8 @@ function Services_Twilio_autoload($className)
         && substr($className, 0, 23) != 'Pricing_Services_Twilio') {
         return false;
     }
-    if(substr($className, 0, 37) == 'Services_Twilio_TaskRouter_Capability' ||
-        substr($className, 0, 45) == 'Services_Twilio_TaskRouter_Worker_Capability' ||
-        substr($className, 0, 48) == 'Services_Twilio_TaskRouter_TaskQueue_Capability' ||
-        substr($className, 0, 48) == 'Services_Twilio_TaskRouter_Workspace_Capability') {
-        $file = "/Twilio/CapabilityTaskRouter";
-    }else {
-        $file = str_replace('_', '/', $className);
-        $file = str_replace('Services/', '', $file);
-    }
+	$file = str_replace('_', '/', $className);
+	$file = str_replace('Services/', '', $file);
     return include dirname(__FILE__) . "/$file.php";
 }
